+++ lib/Auth/Process/IsCesnetEligible.php
@@ -2,10 +2,16 @@
 
 namespace SimpleSAML\Module\cesnet\Auth\Process;
 
+use Jose\Component\Core\AlgorithmManager;
+use Jose\Component\KeyManagement\JWKFactory;
+use Jose\Component\Signature\Algorithm\RS512;
+use Jose\Component\Signature\JWSBuilder;
+use Jose\Component\Signature\Serializer\CompactSerializer;
 use SimpleSAML\Auth\ProcessingFilter;
 use SimpleSAML\Configuration;
 use SimpleSAML\Module;
 use SimpleSAML\Module\perun\Adapter;
+use SimpleSAML\Module\perun\ChallengeManager;
 use SimpleSAML\Module\perun\LdapConnector;
 use SimpleSAML\Module\perun\model\User;
 use SimpleSAML\Module\perun\RpcConnector;
@@ -36,6 +42,9 @@ class IsCesnetEligible extends ProcessingFilter
     const DEFAULT_ATTR_NAME = 'isCesnetEligibleLastSeen';
     const LDAP = 'LDAP';
     const RPC = 'RPC';
+    const SCRIPT_NAME = 'updateIsCesnetEligible';
+    const PATH_TO_KEY = 'pathToKey';
+    const SIGNATURE_ALG = 'signatureAlg';
 
     const PERUN_USER_AFFILIATIONS_ATTR_NAME = 'perunUserAffiliationsAttrName';
     const PERUN_USER_SPONSORING_ORGANIZATIONS_ATTR_NAME = 'perunUserSponsoringOrganizationsAttrName';
@@ -51,6 +60,9 @@ class IsCesnetEligible extends ProcessingFilter
 
     private $idpEntityId;
     private $eduPersonScopedAffiliation = [];
+
+    private $pathToKey;
+    private $signatureAlg;
 
     /**
      * @var LdapConnector
@@ -78,6 +90,12 @@ class IsCesnetEligible extends ProcessingFilter
             );
         }
 
+        if (!isset($config[self::PATH_TO_KEY])) {
+            throw new Exception(
+                'cesnet:isCesnetEligible: missing mandatory configuration option \'pathToKey\'.'
+            );
+        }
+
         if (isset($this->userAffiliationsAttrName, $this->userSponsoringOrganizationsAttrName)) {
             Logger::warning(
                 'cesnet:IsCesnetEligible - One of attributes [' . $this->userAffiliationsAttrName . ', ' .
@@ -86,6 +104,7 @@ class IsCesnetEligible extends ProcessingFilter
         }
 
         $this->rpcAttrName = $config[self::RPC_ATTRIBUTE_NAME];
+        $this->pathToKey = $config[self::PATH_TO_KEY];
 
         $this->cesnetLdapConnector = (new AdapterLdap(self::CONFIG_FILE_NAME))->getConnector();
         $this->rpcAdapter = Adapter::getInstance(Adapter::RPC);
@@ -105,6 +124,12 @@ class IsCesnetEligible extends ProcessingFilter
                 ' is missing or empty. RPC interface will be used'
             );
             $this->adapter =Adapter::getInstance(Adapter::RPC);
+        }
+
+        if (isset($config[self::SIGNATURE_ALG])) {
+            $this->signatureAlg = (array)$config[self::SIGNATURE_ALG];
+        } else {
+            $this->signatureAlg = 'RS512';
         }
 
         $this->userSponsoringOrganizationsAttrName =
@@ -165,15 +190,69 @@ class IsCesnetEligible extends ProcessingFilter
 
             if (!empty($user)) {
                 // Update attribute 'isCesnetEligible' in Perun
-                $data= [
-                    'userId' => $user->getId(),
-                    'isCesnetEligibleValue' => $this->cesnetEligibleLastSeenValue,
-                    'cesnetEligibleLastSeenAttrName' => $this->rpcAttrName
+
+                $id = uniqid();
+
+                $dataChallenge = [
+                    'id' => $id,
+                    'scriptName' => self::SCRIPT_NAME
                 ];
 
-                $cmd = 'curl -X POST -H "Content-Type: application/json" -d \'' . json_encode($data) . '\' ' .
-                       Module::getModuleURL('cesnet/updateIsCesnetEligible.php') . ' > /dev/null &';
-                exec($cmd);
+                $json = json_encode($dataChallenge);
+
+                $curlChallenge = curl_init();
+                curl_setopt($curlChallenge, CURLOPT_POSTFIELDS, $json);
+                curl_setopt($curlChallenge, CURLOPT_URL, Module::getModuleURL('perun/getChallenge.php'));
+                curl_setopt($curlChallenge, CURLOPT_RETURNTRANSFER, true);
+
+                $challenge = curl_exec($curlChallenge);
+
+                if (curl_errno($curlChallenge)) {
+                    $error_msg = curl_error($curlChallenge);
+                    Logger::error('cesnet:IsCesnetEligible - ' . $error_msg);
+                }
+
+                curl_close($curlChallenge);
+
+                if (!empty($challenge)) {
+                    $jwk = JWKFactory::createFromKeyFile($this->pathToKey);
+                    $algorithmManager = new AlgorithmManager(
+                        [
+                            ChallengeManager::getAlgorithm('Signature\\Algorithm', $this->signatureAlg)
+                        ]
+                    );
+                    $jwsBuilder = new JWSBuilder($algorithmManager);
+
+                    $data= [
+                        'userId' => $user->getId(),
+                        'isCesnetEligibleValue' => $this->cesnetEligibleLastSeenValue,
+                        'cesnetEligibleLastSeenAttrName' => $this->rpcAttrName
+                    ];
+
+                    $payload = json_encode([
+                        'iat' => time(),
+                        'nbf' => time(),
+                        'exp' => time() + 3600,
+                        'challenge' => $challenge,
+                        'id' => $id,
+                        'data' => $data
+                    ]);
+
+                    $jws = $jwsBuilder
+                        ->create()
+                        ->withPayload($payload)
+                        ->addSignature($jwk, ['alg' => $this->signatureAlg])
+                        ->build();
+
+                    $serializer = new CompactSerializer();
+                    $token = $serializer->serialize($jws, 0);
+
+                    $cmd = 'curl -X POST -H "Content-Type: application/json" -d \'' . json_encode($token) . '\' ' .
+                        Module::getModuleURL('cesnet/updateIsCesnetEligible.php') . ' > /dev/null &';
+                    exec($cmd);
+                } else {
+                    Logger::error('cesnet:IsCesnetEligible - Retrieving the challenge was not successful.');
+                }
             }
         }